### 2023-11-10 #6050 @editorialbot
[PRE REVIEW]: Wearablehrv: A Python package for the validation of heart rate and heart rate variability in wearables
**Submitting author:** @aminsinichi (<a href="http://orcid.org/0009-0008-2491-1542">Mohammadamin Sinichi</a>)
**Repository:** https://github.com/Aminsinichi/wearable-hrv/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.7
