### 2023-11-11 #6050 @oliviaguest
Hey @Aminsinichi, thanks for choosing JOSS. Can you check the missing DOIs above please and then recompile to check? ☺️
