### 2023-11-17 70ed5ab
feat(import_data): updated it, and also changed chop_data
