### 2023-11-24 #6050 @Aminsinichi
@oliviaguest

Dear Olivia,

I have added a new feature and distributed it via PyPI; therefore, I attempted to upgrade the package version by one step. However, the bot seems to be refusing it. I just wanted to make you aware of this.

Best,
Amin
