### 2023-11-29 1b3c2f4
Update signal_quality function
