### 2023-11-29 35ceb4b
Update signal_quiality function
