### 2023-12-18 29b765b
update group.signal_quality
