### 2023-12-18 18a6293
add feature group.signal_quiality_plot1
