### 2023-12-18 af27c85
add feature group.signal_quality_plot2
