### 2023-12-18 b018dc0
add feature group.mape_analysis
