### 2023-12-18 4f6f245
add feature group.normality_check and group.log_transform_data
