### 2023-12-18 96d395f
update group.regression_plot
