### 2024-01-18 #6240 @editorialbot
[REVIEW]: Wearablehrv: A Python package for the validation of heart rate and heart rate variability in wearables
**Submitting author:** @aminsinichi (<a href="http://orcid.org/0009-0008-2491-1542">Mohammadamin Sinichi</a>)
**Repository:** https://github.com/Aminsinichi/wearable-hrv/
**Branch with paper.md** (empty if default branch):
**Version:** 0.2.0
