### 2024-01-18 #6240 @Aminsinichi
Thank you all for taking the time; very much appreciate it, and looking forward to hearing your thoughts and comments.

@oliviaguest, just a side note (I am not sure how important it is), but version 1.0.10 is currently published (https://pypi.org/project/wearablehrv/).

Thank you!
