### 2024-02-12 Aminsinichi/wearable-hrv#1 @AKuederle
Installation Problem - Use of deprecated astropy API
Using pip install to install the package in a clean env results in errors during import.

Specficially, `ImportError: cannot import name 'LombScargle' from 'astropy.stats'`. This is likely related to the deprecation mentioned here (https://github.com/lightkurve/lightkurve/issues/535)

It seems like the package does not support recent astropy versions. That should be specified in the setup.py to ensure that compatible versions are installed (or even better, the newer versions of astropy should be supported).

Downgrading astropy to <6.0 fixed the issue for now
