### 2024-02-12 Aminsinichi/wearable-hrv#2 @AKuederle
Potentially Confusing naming: CRITERION DEVICE
In the paper and the package the term CRITERION DEVICE is used. At least for me (and I also couldn't find anything with a quick google search), the use of the term in this context was new. I think a more common name would be "reference device" or "ground-truth device"
