### 2024-02-12 Aminsinichi/wearable-hrv#4 @AKuederle
Some comments about Packaging and Versioning
Great job publisheing the package on Pypi!

Here a couple of suggestions to further improve packaging and distribution:

- [x] Your package should have an internal version identifier (usually a __version__ const in the init) that corresponds to the package version, so that users can check which version they have installed
- [x] You should associate each Pypi release with a git tag (and push the tag to Github) to make sure that people can browse the code of exactly the version they are currently running (e.g. when reporting bugs)
- [x] Optionally, you can create a Github release for every version
- [x] Optionally, you can streamline the process, by using Github actions to automatically publish on release or on tag
- [x] You should consider adding a changelog. Otherwise, new versions and their impact are meaningless to users.
