### 2024-02-12 Aminsinichi/wearable-hrv#5 @AKuederle
Tests rely on hardcoded path
At the moment the testfiles try to add hardcoed paths to the sys.path

https://github.com/Aminsinichi/wearable-hrv/blob/ac22c3784c3952a67536c9c976a755827c136dc5/tests/test_group/test_group.py#L3

https://github.com/Aminsinichi/wearable-hrv/blob/ac22c3784c3952a67536c9c976a755827c136dc5/tests/test_individual/test_individual.py#L2

That should be adapted to make it possible to run the tests on every system without modifieing the files
