### 2024-02-12 Aminsinichi/wearable-hrv#6 @AKuederle
Some comments about Docstrings and documentation
- [x] The docstring formatting is inconsistent. Some functions use a numpydoc format while others use a GoogleDoc (?) style format (e.g. https://github.com/Aminsinichi/wearable-hrv/blob/ac22c3784c3952a67536c9c976a755827c136dc5/wearablehrv/individual.py#L1552)
- [x] It would be great, if a rendered API docu (e.g. using RTD) would exist to easier browse the available functions
- [x] The main Documentation example should be restructured, that it can be run **without** modification end-to-end, so that new users can play around with the functionality right away. For this I would suggest replacing, the "undefined" paths with paths to the example file, splitting the docs up into multiple files (1 showing the main functionality and multiple others showing specific functionality, e.g. pre-processing for a specific system)
- [x] Ideally, rendered versions of the example notebook should be hosted somewhere (e.g. as part of RTD page), so that potential users can see the functionality "in action" without installing the package
- [x] If you want to be "fancy" with the handling of your example files, consider using https://pypi.org/project/pooch/ to handle downloading your example files.
- [x] The current documentation notebook has a syntax error in the first cell (missing `!` to mark the first line as command)
- [ ] Ideally add a test that executes your documentation example to make sure that it is working as expected.
