### 2024-02-21 Aminsinichi/wearable-hrv#1 @Aminsinichi
Thank you very much! Addressed the issue in setup.py by defining `atropy<6.0`.
[Related commit]( https://github.com/Aminsinichi/wearable-hrv/commit/09ab91475e91172ebff59f0ef32b6ee9da71de84).
