### 2024-02-21 Aminsinichi/wearable-hrv#2 @Aminsinichi
Thank you very much for your suggestion. I did a quick search through the library of papers to which I had access in my reference manager and found a few papers that used the term "**criterion**" in the context of validation studies. I most likely got inspired by this name after reading one of them. I gathered a few quotes from these 7 papers together with their DOI:

1.	 Data from the 3 devices were time-aligned, and the validity of Polar OH1 and Fitbit Charge 3 was assessed against Polar H10 (**criterion device**) (https://doi.org/10.2196/25313).

2.	Therefore, the purpose of the present study was to test the validity of the Polar® OH1 PPG arm band when compared to a **criterion measure** (Polar® H7 chest strap) during yoga sequences (https://doi.org/10.1055/a-0631-0920).

3.	For each session, participants wore two HR monitors: the Polar OH1 monitor and the Polar 128 H7 belt (Polar Electro Oy, Kempele, Finland) as the **criterion measurement** (https://doi.org/10.1055/a-0875-4033).

4.	Polar OH1 demonstrates high level of agreement with the **criterion measure** ECG HR, thus can be used as a valid measure of HR in lab and field settings during moderate and high intensity physical activities (https://doi.org/10.1371/journal.pone.0217288).

5.	This systematic review and meta-analysis compared the validity of wrist-worn HR estimates to a **criterion measure** of HR (electrocardiography ECG or chest strap)(https://doi.org/10.1080/02640414.2020.1767348).

6.	The Polar H10 appears to be the most accurate wearable device when compared to **criterion measures** and even appears to supersede traditional methods during exercise (https://doi.org/10.3390/s21041061).

7.	Each subject wore a Basis Peak (BPk) on one wrist and a Fitbit Charge HR (FB) on the opposite wrist. **Criterion measurement** of HR was administered by 12-lead ECG (https://pubmed.ncbi.nlm.nih.gov/27803634/).

If you think a better name, as you suggested, is less confusing I will be more than happy to address it.
