### 2024-02-21 c19c010
Add __version__ attribute to __init__.py
