+++ ChANGELOG.md
@@ -1,0 +1,26 @@
+# Changelog
+
+All notable changes to wearablehrv will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+-
+
+### Fixed
+
+-
+
+## [0.1.12] - 2024-02-21
+
+### Added
+
+-
+
+### Fixed
+
+-
