### 2024-02-21 Aminsinichi/wearable-hrv#4 @Aminsinichi
- [x] The first suggestion was about including the __version__ to __init__. Apologies, I accidentally turned this into a new issue and then removed it. Your suggestion has been incorporated [related commit](https://github.com/Aminsinichi/wearable-hrv/commit/c19c010a71203a9a89820452f28785944ca1d455).
- [X] Many thanks for these useful suggestions. I have now added a GitHub Action, which automates the process of publishing on PyPI, and is triggered by pushing the version as a tag, which hopefully will resolve the issue. [Related commit1](https://github.com/Aminsinichi/wearable-hrv/commit/bbce77ff86ff8b622d0b46ab991d832464e0cfdd); [Related commit2](https://github.com/Aminsinichi/wearable-hrv/commit/7340e3ba58fd2724931bb868a9a080839c9c6694)
- [X] Added a template changelog, and will update this with every newer version from now on [Related commit](https://github.com/Aminsinichi/wearable-hrv/commit/81af4cc227cba0ab2e4aeaa79ae72c3d5f513d1b).
