+++ tests/test_group/test_group.py
@@ -1,22 +1,11 @@
-import pdb
-import sys
-sys.path.append(r"C:\Users\msi401\OneDrive - Vrije Universiteit Amsterdam\PhD\Data\Coding\Validation Study\wearable-hrv")
 import unittest
 from wearablehrv import group
 import pandas as pd
-import pingouin as pg
 import os
-import time
-import pickle
-import tkinter as tk
-import hrvanalysis
 from unittest.mock import patch, MagicMock, PropertyMock
 import numpy as np
 import matplotlib.pyplot as plt
-import ipywidgets as widgets
 import seaborn as sns
-from ipywidgets import Dropdown, IntText, Output, HBox
-from IPython.display import display, clear_output
 from copy import deepcopy
 
 path = os.path.dirname(os.path.abspath(__file__)) + "/" 
@@ -138,47 +127,145 @@ class TestGroup(unittest.TestCase):
     
     def test_signal_quality(self):
         # Loading the original dataset
-        sample_data, file_names = group.import_data (path, self.conditions, self.devices, self.features)
+        sample_data, file_names = group.import_data(self.path, self.conditions, self.devices, self.features)
         
-        # Execute the function
-        new_data, new_features = group.signal_quality(
+        # Execute the modified function
+        new_data, new_features, summary_df, quality_df = group.signal_quality(
             deepcopy(sample_data), self.path, self.conditions, self.devices, 
-            self.features, self.criterion, file_names, exclude = True
+            self.features, self.criterion, file_names, exclude=True, 
+            save_as_csv=True, ibi_threshold=0.20, artefact_threshold=0.20, 
+            manual_missing=False, missing_threshold=10
         )
         
-        # Test if the function returns a tuple of (dict, list)
+        # Test if the function returns a tuple of the correct structure
         self.assertIsInstance(new_data, dict)
         self.assertIsInstance(new_features, list)
+        self.assertIsInstance(summary_df, pd.DataFrame)
+        self.assertIsInstance(quality_df, pd.DataFrame)
 
         # Test if 'artefact' and 'nibi_after_cropping' are removed from the features and the data
         self.assertNotIn('artefact', new_features)
         self.assertNotIn('nibi_after_cropping', new_features)
 
-        self.assertNotIn('nibi_after_cropping', new_data["vu"])
-        self.assertNotIn('artefact', new_data["vu"])
+        for device in self.devices:
+            self.assertNotIn('nibi_after_cropping', new_data[device])
+            self.assertNotIn('artefact', new_data[device])
 
-        # Test if the data dictionary is updated correctly 
-        self.assertEqual(sample_data['rhythm']['rmssd']['standing']['P05.csv'], [48.39451400218052]) # before excluding
-        self.assertEqual(new_data['rhythm']['rmssd']['standing']['P05.csv'], []) # after excluding 
- 
-        # Test if the CSV file is created
-        test_file_1 = path + "quality_report1.csv"
-        test_file_2 = path + "quality_report2.csv"
-        self.assertTrue(os.path.exists(test_file_1))
-        self.assertTrue(os.path.exists(test_file_2))
+        # Test the structure of the quality_df DataFrame
+        expected_columns = ['Device', 'Condition', 'Participant', 'Detected Beats', 'Criterion Beats', 'Detected Artefacts', 'Decision']
+        self.assertTrue(all(column in quality_df.columns for column in expected_columns))
 
-        # Test the content of the CSV file
-        saved_data_1 = pd.read_csv(test_file_1)
-        saved_data_2 = pd.read_csv(test_file_2)
-        self.assertEqual(saved_data_1["Decision"].tolist()[81], 'Missing')
-        self.assertEqual(saved_data_2["Acceptable"].tolist()[0:4], [0, 0, 4, 3])
[…]
