### 2024-02-23 Aminsinichi/wearable-hrv#5 @Aminsinichi
Currently the path is dynamically determined based on `path = os.path.dirname(os.path.abspath(__file__))`. Additionally, a few more tests have been added based on the newly added core features. [Related commit](https://github.com/Aminsinichi/wearable-hrv/commit/f75f3f24abfe6510936bd8d33f1f11f47a0f03aa)
