### 2024-02-27 e21f84e
Generate initial Sphinx documentation structure
