+++ docs/source/index.rst
@@ -1,0 +1,24 @@
+.. wearablehrv documentation master file, created by
+   sphinx-quickstart on Tue Feb 27 11:56:41 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to wearablehrv's documentation!
+=======================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+.. automodule:: wearablehrv
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
