### 2024-02-27 110d32e
Update RTD config and add requirements for documentation
