### 2024-02-27 Aminsinichi/wearable-hrv#6 @Aminsinichi
- [x] Updated the docstrings to have a consistent format, [commit1: individual.py](https://github.com/Aminsinichi/wearable-hrv/commit/35a70f1cc141a7971fe5058f0c77cd16e39e162d), [commit2: group.py](https://github.com/Aminsinichi/wearable-hrv/commit/d019ecb94e36bc08d5ba68b9d96dc0e9750f7163)
- [x] Thank you for the very helpful suggestion. I have added API documentation, which is now accessible at [readthedocs](https://wearable-hrv.readthedocs.io/en/latest/). These are the related commits: [commit1](https://github.com/Aminsinichi/wearable-hrv/commit/e21f84e6e33d139534343acfacbafce522d996ca), [commit2](https://github.com/Aminsinichi/wearable-hrv/commit/110d32e125bf65f97ac9f9d94e2d31d6e0916fc7), [commit3](https://github.com/Aminsinichi/wearable-hrv/commit/baf4f4c4f092bd1e9f63f2b232a9dd7be9ab9fce), and I added to the readme.md [commit4](https://github.com/Aminsinichi/wearable-hrv/commit/eff58c1188d4cb2477ba9098aaff0f7c4ea8b8d6).
