### 2024-02-28 17473c6
Update individual pipeline according to PEP8
