+++ wearablehrv/individual.py
@@ -8,7 +8,6 @@
 import datetime
 import os
 import copy
-import json
 import pickle
 import pandas as pd
 import tkinter as tk
@@ -18,7 +17,7 @@ import matplotlib.pyplot as plt
 import ipywidgets as widgets
 import matplotlib.dates as mdates
 from ipywidgets import IntText, Dropdown, Output, HBox
-from IPython.display import display, clear_output, Markdown
+from IPython.display import display, clear_output
 from hrvanalysis import remove_outliers, remove_ectopic_beats, interpolate_nan_values
 from hrvanalysis import get_time_domain_features
 from hrvanalysis import get_frequency_domain_features
@@ -29,13 +28,13 @@ from avro.io import DatumReader
 ###########################################################################
 ###########################################################################
 
-def labfront_conversion (path, pp, file_name, device_name, date):
 
+def labfront_conversion(path, pp, file_name, device_name, date):
     """
     Converts Labfront data into a standardized CSV format, filtering by a specific date.
 
-    This function processes CSV data from Labfront. It first reads the data, then focuses on 
-    the 'isoDate' and 'bbi' columns. The data is then filtered based on the provided date, and 
+    This function processes CSV data from Labfront. It first reads the data, then focuses on
+    the 'isoDate' and 'bbi' columns. The data is then filtered based on the provided date, and
     relevant columns are renamed for standardization. The processed data is then saved into a new CSV file.
 
     Parameters:
@@ -56,29 +55,30 @@ def labfront_conversion (path, pp, file_name, device_name, date):
     None. The function saves the output directly to a CSV file in the specified path.
     """
 
-    labfront = pd.read_csv (path+file_name, skiprows=5)
+    labfront = pd.read_csv(path + file_name, skiprows=5)
     # convert isoDate to datetime column
-    labfront['isoDate'] = pd.to_datetime(labfront['isoDate'])
+    labfront["isoDate"] = pd.to_datetime(labfront["isoDate"])
     # filter rows for the selected date
-    labfront = labfront[labfront['isoDate'].dt.date == pd.to_datetime(date).date()]
-    labfront = labfront [["unixTimestampInMs", "bbi"]]
-    labfront = labfront.rename (columns = {"unixTimestampInMs":"timestamp", "bbi": "rr"})
-    labfront.to_csv (path+pp+"_"+device_name+".csv", index = False)
+    labfront = labfront[labfront["isoDate"].dt.date == pd.to_datetime(date).date()]
+    labfront = labfront[["unixTimestampInMs", "bbi"]]
+    labfront = labfront.rename(columns={"unixTimestampInMs": "timestamp", "bbi": "rr"})
+    labfront.to_csv(path + pp + "_" + device_name + ".csv", index=False)
 
-    print ("Dataset Successfully Converted and Saved in Your Path!")
+    print("Dataset Successfully Converted and Saved in Your Path!")
+
 
 ###########################################################################
 ###########################################################################
 ###########################################################################
 
-def empatica_conversion (path, pp):
-    
+
+def empatica_conversion(path, pp):
     """
     Converts Empatica data from Avro format into a CSV file, focusing on the 'systolicPeaks' field.
 
-    This function processes data files associated with a participant's Empatica device data stored 
-    in Avro format. It specifically reads the 'systolicPeaks' field from these Avro files. 
-    The extracted peak times (in nanoseconds) are converted to milliseconds, and the interbeat 
+    This function processes data files associated with a participant's Empatica device data stored
+    in Avro format. It specifically reads the 'systolicPeaks' field from these Avro files.
+    The extracted peak times (in nanoseconds) are converted to milliseconds, and the interbeat
     intervals (IBIs) are then calculated. The resulting data is saved to a CSV file.
 
     Parameters:
[…]
