### 2024-02-28 d46f36e
Update group pipeline according to PEP8
