+++ wearablehrv/group.py
@@ -11,7 +11,6 @@ import pingouin as pg
 import numpy as np
 from scipy import stats
 import seaborn as sns
-import statsmodels.api as sm
 import plotly.graph_objects as go
 import matplotlib.pyplot as plt
 import matplotlib.transforms as transforms
@@ -25,8 +24,8 @@ from IPython.display import display, clear_output
 ###########################################################################
 ###########################################################################
 
-def import_data (path, conditions, devices, features):
 
+def import_data(path, conditions, devices, features):
     """
     This function imports data from multiple CSV files located in a directory, and structures the data in a dictionary that represents the devices, features, and conditions of the data.
 
@@ -49,7 +48,6 @@ def import_data (path, conditions, devices, features):
         A nested dictionary that represents the imported data, with keys for each device, feature, and condition.
     """
 
-
     # list all files in the directory
     files = os.listdir(path)
     # filter the list to only include CSV files with names matching the pattern "Pxx.csv"
@@ -58,30 +56,45 @@ def import_data (path, conditions, devices, features):
     file_names.sort()
 
     path_files = [os.path.join(path, filename) for filename in file_names]
-    participants = {} #this variable is a dictionary. The keys are each participant's file name, e.g., P01.csv. the values are a big pandas containing the csv file for each participant
+    participants = (
+        {}
+    )  # this variable is a dictionary. The keys are each participant's file name, e.g., P01.csv. the values are a big pandas containing the csv file for each participant
 
     for pp, path_file in zip(file_names, path_files):
         participants[pp] = pd.read_csv(path_file)
 
-    data = {device: {feature: {condition: {pp: [] for pp in file_names} for condition in conditions} for feature in features} for device in devices}
+    data = {
+        device: {
+            feature: {
+                condition: {pp: [] for pp in file_names} for condition in conditions
+            }
+            for feature in features
+        }
+        for device in devices
+    }
 
     for pp in file_names:
         for feature in features:
             for condition in conditions:
                 for device in devices:
-                    value = participants[pp][(participants[pp]["device"] == device) & (participants[pp]["conditions"] == condition)][feature]
+                    value = participants[pp][
+                        (participants[pp]["device"] == device)
+                        & (participants[pp]["conditions"] == condition)
+                    ][feature]
                     data[device][feature][condition][pp].extend(value)
 
-    print ("These are the detected .csv files: {}".format (file_names))
-    print ("Group dataset imported succesfully")
+    print("These are the detected .csv files: {}".format(file_names))
+    print("Group dataset imported succesfully")
 
     return data, file_names
 
+
 ###########################################################################
 ###########################################################################
 ###########################################################################
 
-def nan_handling (data, devices, features, conditions):
+
+def nan_handling(data, devices, features, conditions):
     """
     This function takes a nested dictionary containing the data and replaces [nan] values with an empty list [].
 
@@ -107,19 +120,23 @@ def nan_handling (data, devices, features, conditions):
             for condition in conditions:
                 values = new_data[device][feature][condition]
                 for key, value in values.items():
-                    if isinstance(value, list) and len(value) == 1 and pd.isna(value[0]):
+                    if (
+                        isinstance(value, list)
+                        and len(value) == 1
+                        and pd.isna(value[0])
+                    ):
                         new_data[device][feature][condition][key] = []
-    
-    print ("NaN values are removed from the data")
+
+    print("NaN values are removed from the data")
     return new_data
 
+
 ###########################################################################
[…]
