### 2024-03-11 35a70f1
Improve the consistency of the docstring in Individual.py
