+++ wearablehrv/individual.py
@@ -1,9 +1,4 @@
-#########################INDIVIDUAL######################################
-#########################INDIVIDUAL######################################
-#########################INDIVIDUAL######################################
-############################## Wearablehrv ##############################
-
-############################Importing Modules############################
+# INDIVIDUAL pipeline for Wearablehrv package
 
 import datetime
 import os
@@ -33,26 +28,29 @@ def labfront_conversion(path, pp, file_name, device_name, date):
     """
     Converts Labfront data into a standardized CSV format, filtering by a specific date.
 
-    This function processes CSV data from Labfront. It first reads the data, then focuses on
-    the 'isoDate' and 'bbi' columns. The data is then filtered based on the provided date, and
-    relevant columns are renamed for standardization. The processed data is then saved into a new CSV file.
+    This function reads CSV data from Labfront, focusing on the 'isoDate' and 'bbi' columns. It filters the data based on the provided date and renames relevant columns for standardization. The processed data is saved into a new CSV file named with the participant ID and device name.
 
-    Parameters:
-    -----------
+    Parameters
+    ----------
     path : str
         The directory path pointing to the location of the Labfront data.
     pp : str
         The unique ID of the participant for which the data is being processed.
     file_name : str
-        The name of the Labfront file (with its extension) to be processed.
+        The name of the Labfront file (including its extension) to be processed.
     device_name : str
-        The name of the device used to collect the data. This will be used in the resulting CSV's filename.
+        The name of the device used to collect the data, used in the resulting CSV's filename.
     date : str
-        The specific date for which data should be extracted, provided in a format that can be parsed by pandas' to_datetime function (e.g., 'YYYY-MM-DD').
+        The specific date for which data should be extracted, in a format that can be parsed by pandas' to_datetime function (e.g., 'YYYY-MM-DD').
 
-    Returns:
-    --------
-    None. The function saves the output directly to a CSV file in the specified path.
+    Returns
+    -------
+    None
+        The function saves the output directly to a CSV file in the specified path. The final CSV contains a timestamp column and a renamed 'bbi' column to 'rr', indicating respiratory rate.
+
+    Notes
+    -----
+    The function prints a message upon successful completion of the conversion and saving process.
     """
 
     labfront = pd.read_csv(path + file_name, skiprows=5)
@@ -74,31 +72,33 @@ def labfront_conversion(path, pp, file_name, device_name, date):
 
 def empatica_conversion(path, pp):
     """
-    Converts Empatica data from Avro format into a CSV file, focusing on the 'systolicPeaks' field.
+    Converts Empatica data from Avro format to a CSV file, focusing on the 'systolicPeaks' field.
 
-    This function processes data files associated with a participant's Empatica device data stored
-    in Avro format. It specifically reads the 'systolicPeaks' field from these Avro files.
-    The extracted peak times (in nanoseconds) are converted to milliseconds, and the interbeat
-    intervals (IBIs) are then calculated. The resulting data is saved to a CSV file.
+    This function processes Empatica device data files stored in Avro format for a given participant.
+    It extracts the 'systolicPeaks' field from these files, which includes peak times in nanoseconds.
+    These times are converted to milliseconds, and the interbeat intervals (IBIs) are calculated.
+    The resulting data, comprising timestamps (in milliseconds) and IBIs, is saved to a CSV file.
 
-    Parameters:
-    -----------
+    Parameters
+    ----------
     path : str
         The directory path pointing to the location of the participant's Empatica data.
     pp : str
         The unique ID of the participant whose Empatica data is to be converted.
 
-    Note:
+    Returns
+    -------
+    None
+        The function saves the output directly to a CSV file in the specified path. The CSV file is named
+        using the participant's ID with '_empatica.csv' as the suffix and includes columns for 'timestamp'
[…]
