### 2024-03-18 Aminsinichi/wearable-hrv#2 @AKuederle
Interesting... I have not come across that term eventhough I did a bunch of comparison studies with werables. Usually, the terms "reference system", "gold-standard", or "ground truth" is used. I would suggest adding some some clear explainer to the documentation to ensure that people are not confused by that name.
