### 2024-03-18 Aminsinichi/wearable-hrv#10 @AKuederle
Joss Paper: Comparison to "state-of-the-art" and references for the process
While reading through the paper, I was missing a little bit more thoprough comparison to other available packages to clearer highlight what this package adds. E.g. how much easier is using this package than using the `hrv-analysis` package directly.

Also It would be great, if you could add citations to the paper that provide support for the evaluation metrics that the package provides. Basically, to provide some justification why you chose the set of metrics and plots that are included in the package for comparison of the systems.
