### 2024-03-18 #6240 @AKuederle
@oliviaguest Thanks for the reminder!

I opened some issues in the repo a while back to help the author improve their code, as I had some trouble understanding it/getting things to run:

https://github.com/Aminsinichi/wearable-hrv/issues/5
https://github.com/Aminsinichi/wearable-hrv/issues/4
https://github.com/Aminsinichi/wearable-hrv/issues/6
https://github.com/Aminsinichi/wearable-hrv/issues/1
https://github.com/Aminsinichi/wearable-hrv/issues/2

I was not able to fully validate the funcitonality of the package, as I wanted to wait for the author to respond to all of my suggestions. In particular, I found the example quite hard to understand. I wanted to wait for potential improvement (as suggested in one the issues), to fully work through it.

I read through the paper and it is well written. I was missing the comparison to other papers/packages a little and opened an issue about it (https://github.com/Aminsinichi/wearable-hrv/issues/10)

Regrading general scope, the package primarily provides IO and plotting functionality. The analysis components are all handled by existing packages. The GUI they build around it seems to be nice, but I am not sure if it is good enough and easy enough for people to use it over custom solutions using the analysis packages directly.
