### 2024-03-18 #6240 @Aminsinichi
Thank you very much to both of you! I will continue working through the remaining issues and also make sure to justify the added value of the package in the paper as well. I will also improve the documentation and examples as you suggested, which should hopefully help you to test the full capabilities.
