### 2024-03-18 d019ecb
Improve the consistency of the docstring in Group.py
