+++ wearablehrv/group.py
@@ -1,9 +1,4 @@
-##################################GROUP####################################
-##################################GROUP####################################
-##################################GROUP####################################
-############################### Wearablehrv ###############################
-
-############################Importing Modules############################
+# Group pipeline for Wearablehrv package
 
 import os
 import pandas as pd
@@ -27,27 +22,32 @@ from IPython.display import display, clear_output
 
 def import_data(path, conditions, devices, features):
     """
-    This function imports data from multiple CSV files located in a directory, and structures the data in a dictionary that represents the devices, features, and conditions of the data.
+    Imports data from multiple CSV files located in a specific directory and structures the data into a nested dictionary based on devices, features, and conditions.
 
-    Parameters:
-    -----------
-    file_names : list
-        A list of strings containing the file names of the CSV files to be imported.
+    This function scans a directory for CSV files, reads the data from these files, and organizes it into a nested dictionary. This dictionary is structured to allow easy access to data based on device, feature, and condition. Each file is assumed to contain data for all conditions, devices, and features specified by the parameters.
+
+    Parameters
+    ----------
     path : str
-        The path to the directory where the CSV files are located.
-    conditions : list
-        A list of strings representing the different experimental conditions of the data.
-    devices : list
+        The path to the directory where the CSV files are located. The function expects to find files named in a specific pattern that includes the participant ID.
+    conditions : list of str
+        A list of strings representing the different experimental conditions under which the data was collected.
+    devices : list of str
         A list of strings representing the different devices used to collect the data.
-    features : list
-        A list of strings representing the different features of the data.
+    features : list of str
+        A list of strings representing the different features recorded in the data.
 
-    Returns:
-    --------
-    data : dict
-        A nested dictionary that represents the imported data, with keys for each device, feature, and condition.
+    Returns
+    -------
+    tuple
+        A tuple containing two elements:
+        - data: A nested dictionary with the structure {device: {feature: {condition: {participant: [values]}}}}, representing the imported data.
+        - file_names: A list of the file names that were imported, providing a record of which files were processed.
+
+    Notes
+    -----
+    - The function automatically identifies CSV files in the provided path and imports data from them. The expected file naming convention and data structure within the files should match the specified devices, conditions, and features.
     """
-
     # list all files in the directory
     files = os.listdir(path)
     # filter the list to only include CSV files with names matching the pattern "Pxx.csv"
@@ -138,26 +138,29 @@ def nan_handling(data, devices, features, conditions):
 
 def save_data(data, path, conditions, devices, features, file_names):
     """
-    This function takes in a dictionary of data, along with information about the experimental conditions, devices, and features, and saves the data to a CSV file.
+    Saves aggregated data from multiple sources to a CSV file based on specified conditions, devices, and features.
 
-    Parameters:
-    -----------
+    This function compiles data into a single DataFrame and saves it to a CSV file. The data is organized by combining specified conditions, devices, and features into unique column names.
+
+    Parameters
+    ----------
     data : dict
-        A nested dictionary that represents the data to be saved, with keys for each device, feature, and condition.
+        A nested dictionary with structure {device: {feature: {condition: [values]}}}, representing the data to be saved.
     path : str
-        The path to the directory where the CSV file should be saved.
-    conditions : list
-        A list of strings representing the different experimental conditions of the data.
-    devices : list
-        A list of strings representing the different devices used to collect the data.
-    features : list
-        A list of strings representing the different features of the data.
-    file_names : list
-        A list of strings containing the file names of the participants.
[…]
