### 2024-03-19 f3a8ad8
replace path (str) with Path from pathlib
