+++ wearablehrv/individual.py
@@ -18,6 +18,7 @@ from hrvanalysis import get_time_domain_features
 from hrvanalysis import get_frequency_domain_features
 from avro.datafile import DataFileReader
 from avro.io import DatumReader
+from pathlib import Path
 
 ###########################################################################
 ###########################################################################
@@ -32,8 +33,8 @@ def labfront_conversion(path, pp, file_name, device_name, date):
 
     Parameters
     ----------
-    path : str
-        The directory path pointing to the location of the Labfront data.
+    path : str or Path
+        The directory path pointing to the location of the Labfront data. Can be a string or a Path object.
     pp : str
         The unique ID of the participant for which the data is being processed.
     file_name : str
@@ -53,14 +54,17 @@ def labfront_conversion(path, pp, file_name, device_name, date):
     The function prints a message upon successful completion of the conversion and saving process.
     """
 
-    labfront = pd.read_csv(path + file_name, skiprows=5)
+    path = Path(path)
+    labfront_file_path = path / file_name
+    output_file_path = path / f"{pp}_{device_name}.csv"
+    labfront = pd.read_csv(labfront_file_path + file_name, skiprows=5)
     # convert isoDate to datetime column
     labfront["isoDate"] = pd.to_datetime(labfront["isoDate"])
     # filter rows for the selected date
     labfront = labfront[labfront["isoDate"].dt.date == pd.to_datetime(date).date()]
     labfront = labfront[["unixTimestampInMs", "bbi"]]
     labfront = labfront.rename(columns={"unixTimestampInMs": "timestamp", "bbi": "rr"})
-    labfront.to_csv(path + pp + "_" + device_name + ".csv", index=False)
+    labfront.to_csv(output_file_path, index=False)
 
     print("Dataset Successfully Converted and Saved in Your Path!")
 
@@ -81,8 +85,9 @@ def empatica_conversion(path, pp):
 
     Parameters
     ----------
-    path : str
+    path : str or Path
         The directory path pointing to the location of the participant's Empatica data.
+        This can be provided as a string or a Path object from the pathlib module.
     pp : str
         The unique ID of the participant whose Empatica data is to be converted.
 
@@ -99,7 +104,8 @@ def empatica_conversion(path, pp):
     - The function assumes Avro files contain the 'systolicPeaks' field with peak times in nanoseconds.
     - The final CSV file excludes the last timestamp since there's no corresponding IBI.
     """
-    avrofiles_path = path + "/" + pp + "_empatica" + "/raw_data" + "/v6"
+    path = Path(path)
+    avrofiles_path = path / f"{pp}_empatica/raw_data/v6"
 
     # Function to read systolicPeaks data from a single Avro file
     def read_systolic_peaks_from_file(avro_file):
@@ -114,11 +120,11 @@ def empatica_conversion(path, pp):
 
     # Iterate through all Avro files in the given directory and combine the 'systolicPeaks' data
     combined_systolic_peaks_ns = []
-    for file in os.listdir(avrofiles_path):
-        if file.endswith(".avro"):
-            avro_file = os.path.join(avrofiles_path, file)
-            systolic_peaks = read_systolic_peaks_from_file(avro_file)
-            combined_systolic_peaks_ns.extend(systolic_peaks["peaksTimeNanos"])
+    for avro_file in avrofiles_path.glob(
+        "*.avro"
+    ):  # Using Path.glob to find .avro files
+        systolic_peaks = read_systolic_peaks_from_file(avro_file)
+        combined_systolic_peaks_ns.extend(systolic_peaks["peaksTimeNanos"])
 
     # convert each nanosecond value to millisecond
[…]
