+++ wearablehrv/group.py
@@ -14,6 +14,7 @@ import ipywidgets as widgets
 from scipy.stats import linregress
 from copy import deepcopy
 from IPython.display import display, clear_output
+from pathlib import Path
 
 ###########################################################################
 ###########################################################################
@@ -48,20 +49,18 @@ def import_data(path, conditions, devices, features):
     -----
     - The function automatically identifies CSV files in the provided path and imports data from them. The expected file naming convention and data structure within the files should match the specified devices, conditions, and features.
     """
+    path = Path(path)
     # list all files in the directory
-    files = os.listdir(path)
-    # filter the list to only include CSV files with names matching the pattern "Pxx.csv"
-    file_names = [f for f in files if f.endswith(".csv")]
+    file_names = [f.name for f in path.glob("*.csv") if f.is_file()]
     # sort the resulting list of file names
     file_names.sort()
 
-    path_files = [os.path.join(path, filename) for filename in file_names]
     participants = (
         {}
     )  # this variable is a dictionary. The keys are each participant's file name, e.g., P01.csv. the values are a big pandas containing the csv file for each participant
 
-    for pp, path_file in zip(file_names, path_files):
-        participants[pp] = pd.read_csv(path_file)
+    for pp in file_names:
+        participants[pp] = pd.read_csv(path / pp)
 
     data = {
         device: {
@@ -146,8 +145,8 @@ def save_data(data, path, conditions, devices, features, file_names):
     ----------
     data : dict
         A nested dictionary with structure {device: {feature: {condition: [values]}}}, representing the data to be saved.
-    path : str
-        The directory path where the CSV file will be saved.
+    path : str or Path
+        The directory path where the CSV file will be saved. Can be provided as a string or a Path object from the pathlib module.
     conditions : list of str
         Experimental conditions under which the data was collected.
     devices : list of str
@@ -176,7 +175,8 @@ def save_data(data, path, conditions, devices, features, file_names):
                 group_data[column_name] = column
 
     # Save the datafile:
-    path_save = path + "group_data.csv"  # creating the path
+    path = Path(path)
+    path_save = path / "group_data.csv"
 
     group_data.to_csv(path_save, index=False)
     print("Data Saved Succesfully!")
@@ -212,8 +212,8 @@ def signal_quality(
     data : dict
         A nested dictionary containing the data to be analyzed. Keys represent each device, feature, and condition.
 
-    path : str
-        The directory path where the quality report CSV files should be saved.
+    path : str or Path
+        The directory path where the quality report CSV files should be saved. Can be provided as a string or a Path object from the pathlib module.
 
     conditions : list
         A list of strings indicating the different experimental conditions in the data.
@@ -348,8 +348,9 @@ def signal_quality(
 
     # Save the datafiles:
     if save_as_csv == True:
-        path_save_quality = path + "quality_report1.csv"  # creating the path
-        path_save_summary = path + "quality_report2.csv"
+        path = Path(path)
+        path_save_quality = path / "quality_report1.csv"
+        path_save_summary = path / "quality_report2.csv"
 
         quality_df.to_csv(path_save_quality, index=False)
         summary_df.to_csv(path_save_summary)
@@ -487,9 +488,7 @@ def signal_quality_plot1(
 ###########################################################################
 
 
-def signal_quality_plot2(
-    summary_df, conditions, devices, condition_selection=False, condition=None
-):
+def signal_quality_plot2(summary_df, condition_selection=False, condition=None):
     """
     Generates a stacked bar plot to visualize data quality by device. This function allows the option to either
     display data for all conditions or focus on a single specified condition. It aggregates data quality metrics
@@ -754,7 +753,7 @@ def box_plot(data, conditions, features, devices):
[…]
