### 2024-03-19 da41ec5
add data.py which incorporates pooch to download the data
