+++ wearablehrv/__init__.py
@@ -2,7 +2,8 @@
 #### A Python package designed for data preparation, pre-processing, feature exctraction, comparison, visualization, and statistical individual/group analysis of heart rate and heart rate variability outcome variables recorded from wearable devices that transmit raw interbeat intervals to a host device via Bluetooth.
 ############################ Wearablehrv ############################
 
-__version__ = '0.1.12'
+__version__ = "0.1.12"
 
-from wearablehrv import individual
-from wearablehrv import group
+from . import individual
+from . import group
+from . import data
