+++ docs/requirements.txt
@@ -2,6 +2,7 @@ sphinx>=4.0
 pandas
 numpy
 plotly
+pooch
 matplotlib
 ipywidgets
 IPython
