### 2024-03-19 Aminsinichi/wearable-hrv#6 @Aminsinichi
- [x] I added `data.py` to the main root that has a few functions and uses `pooch`. This will help automatically download the example data, and extract their Path for immediate execution of all the functions; [related commit](https://github.com/Aminsinichi/wearable-hrv/commits/master/).
