+++ README.md
@@ -1,3 +1,4 @@
+![cover](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/img/cover.png)
 
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![Current version at PyPI](https://img.shields.io/pypi/v/wearablehrv.svg)](https://pypi.org/project/wearablehrv/)
@@ -5,97 +6,35 @@
 ![Last Commit](https://img.shields.io/github/last-commit/Aminsinichi/wearable-hrv)
 [![Twitter Follow](https://img.shields.io/twitter/follow/AminSinichi.svg?style=social)](https://twitter.com/AminSinichi)
 
-`wearablehrv` is an open-source Python package tailored for data preparation, pre-processing, feature extraction, comparison, visualization, and both individual and group statistical analyses of heart rate and heart rate variability metrics from wearable devices that transmit raw inter-beat intervals and timestamps. The inclusion of graphical user interfaces in most functions grants researchers the flexibility to easily switch between experimental conditions and devices. This offers versatility in validating an unlimited number of wearables within a single experimental setting and under various conditions. The only required inputs for the pipeline are the inter-beat intervals and timestamps for each device; the rest is fully handled by the `wearablehrv` package. The main functionalities of this Python package are:
-
-![Image Description](https://showme.redstarplugin.com/d/d:N6ru0hU4)
-
-**Individual Pipeline**:
-
-1. Define experimental events by importing raw data from an unlimited number of devices and experimental conditions.
-2. Visualize the inter-beat-interval against the criterion device using an intuitive graphical user interface.
-3. Correct for lag between devices with millisecond precision and crop the signals when necessary.
-4. Pre-process and calculate both time-domain and frequency-domain measures in one go for all devices and conditions.
-5. Provide various plotting options to compare the criterion devices with a specific device and establish its validity.
-
-**Group Pipeline**:
-
-1. Import all individual cases, perform an extensive signal quality check and analysis, and exclude outliers if necessary based on modifiable cutoffs.
-2. Offer many descriptive plots to visualize the entirety of data for all cases, conditions, and devices.
-3. Conduct the most important statistical analyses in one go and with one click for all devices and conditions against the criterion device, including regression analysis, intraclass correlation coefficient, and Bland-Altman analysis.
+`wearablehrv` is a Python package that comes in handy if you want to validate wearables and establish their accuracy in terms of heart rate (HR) and heart rate variability (HRV). `wearablehrv` is a complete and comprehensive pipeline that helps you go from your recorded raw data through all the necessary pre-processing steps, data analysis, and many visualization tools with graphical user interfaces.
 
 # Documentation
-
-For an in-depth explanation of the package and sample data, please refer to:
-
-[![Documentation](https://img.shields.io/badge/Read-Documentation-blue)](https://github.com/Aminsinichi/wearable-hrv/blob/master/notebooks/documentation.ipynb)
 
 For the complete documentation of the API and modules, visit:
 
 [![Documentation Status](https://readthedocs.org/projects/wearable-hrv/badge/?version=latest)](https://wearable-hrv.readthedocs.io/en/latest/?badge=latest)
 
-# Questions
+**Examples**
 
[…]
