+++ docs/examples/group_pipeline/README.md
@@ -1,0 +1,17 @@
+# wearablehrv - Group Pipeline Guide
+
+Welcome to the Group Pipeline guide for the `wearablehrv` package. This guide will help you navigate through the example notebooks designed for analyzing data from multiple participants in a group setting. The Group Pipeline facilitates statistical analyses and visualizations to evaluate the performance and agreement of wearable devices against a criterion device across various conditions. 
+
+## Overview
+
+1. **group_data_preparation.ipynb**
+   - Instructions on preparing your dataset for group analysis. This includes organizing participant data files and ensuring they are correctly formatted after individual analysis.
+   
+2. **group_signal_quality.ipynb**
+   - A comprehensive guide to assessing signal quality across devices and conditions. Learn how to generate reports and visualizations that highlight the signal quality of your devices.
+
+3. **group_data_analysis.ipynb**
+   - Dive into statistical analyses including Mean Absolute Percentage Error (MAPE), Regression Analysis, Intraclass Correlation Coefficient (ICC), and Bland-Altman Analysis. This notebook will guide you through validating the performance of each device against the criterion device.
+
+4. **group_data_plotting.ipynb**
+   - A variety of plotting options to visualize your group data. This includes violin plots, box plots, radar plots, histogram plots, bar plots, and matrix plots.
