+++ docs/examples/getting_started/README.md
@@ -1,0 +1,11 @@
+# wearablehrv - Getting Started
+
+This guide provides an overview of the entire process, from installation to advanced data analysis and plotting. 
+
+## Overview
+
+- **installation.ipynb**
+  - A step-by-step guide to installing `wearablehrv` and setting up your environment. Ensure you meet the prerequisites and follow the instructions for a smooth installation process.
+
+- **overview.ipynb**
+  - Provides a broad overview of the functionalities and features offered by `wearablehrv`. Explore the capabilities of the package and get acquainted with its components.
