+++ README.md
@@ -23,18 +23,18 @@ For the complete documentation of the API and modules, visit:
 
 *Individual Pipeline*
 
-- [How to prepare your data for the individual pipeline](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/1.individual_data_preparation.ipynb)
-- [Preprocess your data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/2.individual_data_preprocessing.ipynb)
-- [Analyze your data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/3.individual_data_analysis.ipynb)
-- [Plot your data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/4.individual_data_plotting.ipynb)
-- [Learn more about the compatibility of wearablehrv with other platforms (Labfront, VU-AMS, Empatica)](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/individual_compatibility.ipynb)
+- [How to prepare your data for the individual pipeline](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/1.individual_data_preparation.ipynb)
+- [Preprocess your data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/2.individual_data_preprocessing.ipynb)
+- [Analyze your data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/3.individual_data_analysis.ipynb)
+- [Plot your data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/4.individual_data_plotting.ipynb)
+- [Learn more about the compatibility of wearablehrv with other platforms (Labfront, VU-AMS, Empatica)](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/individual_compatibility.ipynb)
 
 *Group Pipeline*
 
-- [How to prepare your data for the group pipeline](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/1.group_data_preparation.ipynb)
-- [Determine the signal quality of your wearables](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/2.group_signal_quality.ipynb)
-- [Perform four major statistical analyses to determine validity](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/3.group_data_analysis.ipynb)
-- [Descriptive plots for your group data](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/4.group_data_plotting.ipynb)
+- [How to prepare your data for the group pipeline](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/1.group_data_preparation.ipynb)
+- [Determine the signal quality of your wearables](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/2.group_signal_quality.ipynb)
+- [Perform four major statistical analyses to determine validity](https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/3.group_data_analysis.ipynb)
[…]
