+++ docs/source/conf.py
@@ -5,28 +5,34 @@
 
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../..'))
+
+sys.path.insert(0, os.path.abspath("../.."))
 
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'wearablehrv'
-copyright = '2024, Amin Sinichi'
-author = 'Amin Sinichi'
-release = '0.1.12'
+project = "wearablehrv"
+copyright = "2024, Amin Sinichi"
+author = "Amin Sinichi"
+release = "0.1.12"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ['sphinx.ext.autodoc']
+extensions = [
+    "sphinx.ext.autodoc",
+    "recommonmark",
+    "sphinx_markdown_tables",
+    "sphinx_rtd_theme",
+]
 
-templates_path = ['_templates']
+
+templates_path = ["_templates"]
 exclude_patterns = []
-
 
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = 'alabaster'
-html_static_path = ['_static']
+html_theme = "sphinx_rtd_theme"
+html_static_path = ["_static"]
