+++ docs/requirements.txt
@@ -13,3 +13,6 @@ seaborn
 statsmodels
 scipy
 astropy<6.0
+sphinx_markdown_tables
+recommonmark
+sphinx_rtd_theme
