+++ docs/source/index.rst
@@ -3,8 +3,7 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to wearablehrv's documentation!
-=======================================
+.. mdinclude:: ../../README.md
 
 .. toctree::
    :maxdepth: 2
