+++ docs/source/conf.py
@@ -21,7 +21,7 @@ release = "0.1.12"
 
 extensions = [
     "sphinx.ext.autodoc",
-    "recommonmark",
+    "m2r2",
     "sphinx_markdown_tables",
     "sphinx_rtd_theme",
 ]
