+++ docs/source/api_individual_pipeline.rst
@@ -1,0 +1,7 @@
+API - Individual Pipeline
+=========================
+
+.. automodule:: wearablehrv.individual
+   :members:
+   :undoc-members:
+   :show-inheritance:
