+++ docs/source/api_group_pipeline.rst
@@ -1,0 +1,7 @@
+API - Group Pipeline
+====================
+
+.. automodule:: wearablehrv.group
+   :members:
+   :undoc-members:
+   :show-inheritance:
