+++ docs/source/index.rst
@@ -3,21 +3,17 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
+Welcome to the wearablehrv documentation!
+=========================================
+
 .. mdinclude:: ../../README.md
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
-.. automodule:: wearablehrv.individual
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. automodule:: wearablehrv.group
-   :members:
-   :undoc-members:
-   :show-inheritance:
+   api_individual_pipeline
+   api_group_pipeline
 
 Indices and tables
 ==================
@@ -25,3 +21,4 @@ Indices and tables
 * :ref:`genindex`
 * :ref:`modindex`
 * :ref:`search`
+
