+++ docs/source/overview.rst
@@ -1,0 +1,193 @@
+Overview
+========
+
+The package is divided into two broad ranges of functionalities:
+
+- **Individual Pipeline**: You use it for a single participant to process your raw data.
+- **Group Pipeline**: You use it when you have multiple participants, and you have processed them through the Individual Pipeline.
+
+Below, we offer a quick overview of the main functionalities.
+
+Data Collection
+---------------
+
+When one wants to establish the validity of a wearable, let's say a smartwatch, that records heart rate and heart rate variability, they should use a "ground truth" device. This is usually a gold-standard electrocardiography (ECG) that measures HR and HRV accurately.
+
+**Note**: We call this gold-standard a "criterion" device in our pipeline.
+
+Then, a participant wears this ECG, together with the smartwatch, and starts recording data simultaneously. It is beneficial if we test the subject in various conditions, so we get a better sense of how well the device works.
+
+Usually, validating multiple devices at once is a cumbersome task, requiring a lot of data preparation, processing, different alignments, etc. **A powerful feature in ``wearablehrv`` is that it does not matter how many devices in how many conditions you want to test a participant!** You just record your data, and the pipeline walks you through this data to the final decision on whether a device is accurate compared to the ground truth or not.
+
+This is how your experiment may look like: a participant wearing a few wearables named Kyto, Heartmath, Empatica, Rhythm, together with a gold-standard ECG (VU-AMS), with electrodes on the chest, and will perform different tasks in different conditions (e.g., sitting for 5 minutes, standing up for 3 minutes, walking for 3 minutes, and biking for 3 minutes, while having all the devices on):
+
+.. image:: https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/img/Sensor%20Placement.png?raw=true
+    :alt: Sensor Placement
+
+Individual Pipeline
+-------------------
+
+Prepare Data
+^^^^^^^^^^^^
+
+It is easy to read your data and experimental events with the pipeline from all your devices in one go.
+
+.. code-block:: python
+
+    # Importing Module
+    import wearablehrv
+
+    # downloading some example data
+    path = wearablehrv.data.download_data_and_get_path()
+    # Define the participant ID 
+    pp = "test" 
+    # Define your experimental conditions, for instance, sitting, standing, walking, and biking
+    conditions = ['sitting', 'standing', 'walking', 'biking'] 
+
+    # Define the devices you want to validate against the criterion. 
+    devices = ["kyto", "heartmath", "rhythm", "empatica", "vu"] 
+
+    # Redefine the name of the criterion device
+    criterion = "vu" 
+
+    # Read data, experimental events, and segment the continuous data into smaller chunks
+    data = wearablehrv.individual.import_data (path, pp, devices)
+    events = wearablehrv.individual.define_events (path, pp, conditions, already_saved= True, save_as_csv= False)
+    data_chopped = wearablehrv.individual.chop_data (data, conditions, events, devices)
+
+Preprocess Data
+^^^^^^^^^^^^^^^
+
+You have various methods to properly preprocess your raw data.
+
+**Correct the Lag, Trim Data**
+
+With a user-friendly GUI, correct the lag between devices, align data by cropping the beginning and the end of each of your devices, and have full control over each device and condition.
+
+.. code-block:: python
+
+    wearablehrv.individual.visual_inspection (data_chopped, devices, conditions,criterion)
+
[…]
