+++ docs/source/index.rst
@@ -6,12 +6,43 @@
 Welcome to the wearablehrv documentation!
 =========================================
 
-.. mdinclude:: ../../README.md
+.. image:: https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/img/cover.png
+    :alt: cover
+
+.. image:: https://img.shields.io/badge/License-MIT-yellow.svg
+    :target: https://opensource.org/licenses/MIT
+    :alt: License: MIT
+
+.. image:: https://img.shields.io/pypi/v/wearablehrv.svg
+    :target: https://pypi.org/project/wearablehrv/
+    :alt: Current version at PyPI
+
+.. image:: https://img.shields.io/pypi/pyversions/wearablehrv.svg
+    :alt: Supported Python Versions
+
+.. image:: https://img.shields.io/github/last-commit/Aminsinichi/wearable-hrv
+    :alt: Last Commit
+
+.. image:: https://mybinder.org/badge_logo.svg
+    :target: https://mybinder.org/v2/gh/Aminsinichi/wearable-hrv/master?labpath=docs%2Fexamples%2F
+    :alt: Binder
+
+``wearablehrv`` is a Python package that comes in handy if you want to validate wearables and establish their accuracy in terms of heart rate (HR) and heart rate variability (HRV). ``wearablehrv`` is a complete and comprehensive pipeline that helps you go from your recorded raw data through all the necessary pre-processing steps, data analysis, and many visualization tools with graphical user interfaces.
+
+Questions
+=========
+
+For any questions regarding the package, please contact:
+
+- `aminsinichi@gmail.com <mailto:aminsinichi@gmail.com>`
+- `m.sinichi@vu.nl <mailto:m.sinichi@vu.nl>`
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
+   overview
+   tutorial
    api_individual_pipeline
    api_group_pipeline
 
