+++ docs/requirements.txt
@@ -14,5 +14,4 @@ statsmodels
 scipy
 astropy<6.0
 sphinx_markdown_tables
-m2r2
 sphinx_rtd_theme
