+++ docs/source/tutorial.rst
@@ -1,0 +1,31 @@
+Tutorial - Jupyter Notebook Examples
+=====================================
+
+Getting Started
+---------------
+
+- `Installation guide for wearablehrv <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/installation.ipynb>`_
+- `An overview of main functionalities <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/getting_started/overview.ipynb>`_
+
+Individual Pipeline
+-------------------
+
+- `How to prepare your data for the individual pipeline <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/1.individual_data_preparation.ipynb>`_
+- `Preprocess your data <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/2.individual_data_preprocessing.ipynb>`_
+- `Analyze your data <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/3.individual_data_analysis.ipynb>`_
+- `Plot your data <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/4.individual_data_plotting.ipynb>`_
+- `Learn more about the compatibility of wearablehrv with other platforms (Labfront, VU-AMS, Empatica) <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/individual_pipeline/individual_compatibility.ipynb>`_
+
+Group Pipeline
+--------------
+
+- `How to prepare your data for the group pipeline <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/1.group_data_preparation.ipynb>`_
+- `Determine the signal quality of your wearables <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/2.group_signal_quality.ipynb>`_
+- `Perform four major statistical analyses to determine validity <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/3.group_data_analysis.ipynb>`_
+- `Descriptive plots for your group data <https://github.com/Aminsinichi/wearable-hrv/blob/master/docs/examples/group_pipeline/4.group_data_plotting.ipynb>`_
+
+You can also explore the example notebooks directly in your browser without installing any packages by using Binder. Simply click the badge below to get started:
+
+.. image:: https://mybinder.org/badge_logo.svg
+    :target: https://mybinder.org/v2/gh/Aminsinichi/wearable-hrv/master?labpath=docs%2Fexamples%2F
+    :alt: Binder
