### 2024-03-22 fe0b23b
update .readthedocs.yaml to only create html
