+++ .readthedocs.yaml
@@ -16,7 +16,8 @@ build:
     python: "3.7"  
 
 # Optionally specify the formats to build
-formats: all
+formats: 
+  - html # Only build HTML documentation
 
 # Define the Python environment and dependencies
 python:
@@ -24,4 +25,5 @@ python:
     - method: pip
       path: .
     - requirements: docs/requirements.txt
+
 
