### 2024-03-22 efece35
update sphinx links to include images
