+++ docs/source/overview.rst
@@ -21,7 +21,7 @@ Usually, validating multiple devices at once is a cumbersome task, requiring a l
 
 This is how your experiment may look like: a participant wearing a few wearables named Kyto, Heartmath, Empatica, Rhythm, together with a gold-standard ECG (VU-AMS), with electrodes on the chest, and will perform different tasks in different conditions (e.g., sitting for 5 minutes, standing up for 3 minutes, walking for 3 minutes, and biking for 3 minutes, while having all the devices on):
 
-.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/blob/master/docs/img/Sensor%20Placement.png?raw=true
+.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/master/docs/img/Sensor%20Placement.png?raw=true
     :alt: Sensor Placement
 
 Individual Pipeline
@@ -68,7 +68,7 @@ With a user-friendly GUI, correct the lag between devices, align data by croppin
 
     wearablehrv.individual.visual_inspection (data_chopped, devices, conditions,criterion)
 
-.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/blob/master/docs/img/visual_inspection.PNG
+.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/master/docs/img/visual_inspection.PNG
     :alt: visual_inspection
 
 **Detect Outliers and Ectopic Beats**
@@ -87,7 +87,7 @@ Check how well you performed the preprocessing by comparing the detected outlier
 
     wearablehrv.individual.ibi_comparison_plot(data_chopped, data_pp, devices, conditions, criterion, width=20, height=10)
 
-.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/blob/master/docs/img/comparison_plot.PNG
+.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/master/docs/img/comparison_plot.PNG
     :alt: comparison_plot
 
 Analyze and Plot
@@ -100,7 +100,7 @@ Easily calculate all relevant outcome variables (e.g., RMSSD, mean HR, frequency
     time_domain_features, frequency_domain_features = wearablehrv.individual.data_analysis(data_pp, devices, conditions)
     wearablehrv.individual.bar_plot(time_domain_features, frequency_domain_features, devices, conditions, width=20, height=25, bar_width = 0.15)
 
-.. image:: https://raw.githubusercontent/Aminsinichi/wearable-hrv/blob/master/docs/img/bar_plot.PNG
+.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/master/docs/img/bar_plot.PNG
     :alt: bar_plot
 
 Group Pipeline
@@ -132,7 +132,7 @@ A powerful tool to assess and report signal quality in all your wearables, in al
     data, features, summary_df, quality_df = wearablehrv.group.signal_quality(data, path, conditions, devices, features, criterion, file_names, ibi_threshold = 0.30, artefact_threshold = 0.30)
     wearablehrv.group.signal_quality_plot2(summary_df, condition_selection=False, condition=None)
 
-.. image:: https://raw.githubusercontent/Aminsinichi/wearable-hrv/blob/master/docs/img/signal_quality.PNG
+.. image:: https://raw.githubusercontent.com/Aminsinichi/wearable-hrv/master/docs/img/signal_quality.PNG
     :alt: signal_quality
 
 Statistical Analysis
@@ -147,7 +147,7 @@ Perform four of the most common statistical methods for validation, and create p
     mape_data = wearablehrv.group.mape_analysis(data, criterion, devices, conditions, features)
[…]
