### 2024-03-22 c0a84ac
update sphinx links to include images
