+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+# Contributing to wearablehrv
+
+## Ways to Contribute
+
+There are many ways you can contribute to `wearablehrv`:
+
+### Reporting Bugs
+
+If you find a bug in the software, please check the existing [issues](https://github.com/Aminsinichi/wearable-hrv/issues) to see if it has already been reported. If it hasn't, submit a new issue with as much detail as possible, including:
+
+- A concise description of the problem
+- Steps to reproduce the issue
+- Expected behavior
+- Actual behavior
+- Your environment details (OS, Python version, etc.)
+
+### Suggesting Enhancements
+
+Have an idea for a new feature or think something could be done better? Please submit an issue with your enhancement proposal.
+
+### Pull Requests
+
+Ready to contribute code or documentation? Here's how you can submit a pull request (PR):
+
+1. **Fork the repository** and clone it locally.
+2. **Create a new branch** for your contribution.
+3. **Make your changes.** Be sure to adhere to the project's coding standards.
+4. **Test your changes.** Ensure that your changes do not break any existing functionality.
+5. **Submit a pull request.** Include a clear description of your changes and any relevant issue numbers.
+
+## Support
+
+Looking for help with `wearablehrv`? Please [submit an issue](https://github.com/Aminsinichi/wearable-hrv/issues) describing your question or problem. You can also drop me an email: [aminsinichi@gmail.com](mailto:aminsinichi@gmail.com).
