### 2024-03-22 Aminsinichi/wearable-hrv#6 @Aminsinichi
Thank you very much for these suggestions. I made many changes based on your comments, summarized below:
- I removed the previous documentation and completely changed it. I have now created several .ipynb examples that show different functionalities step by step. There is also an overview notebook that shows some of the main functionalities of the package. These are all now listed in the README.md, and are located at [docs/examples](https://github.com/Aminsinichi/wearable-hrv/tree/master/docs/examples); [Related commit](https://github.com/Aminsinichi/wearable-hrv/commit/9a4b4d76cfefdae890aec25b767fe2679823cfc9).
- Thanks to `pooch`, users can now download sample data into their `Cache`, and then I retrieve the `Path` and read those files into the pipeline. So, they can now explore all the functions without modifying any code.
- I completely changed the README.md file. It now contains the output images of the “overview.ipynb” notebook, so users can see what it may look like before installing it.
- I also added [Binder](https://mybinder.org/v2/gh/Aminsinichi/wearable-hrv/master?labpath=docs%2Fexamples%2F), so they can try the notebook examples; [Related commit](https://github.com/Aminsinichi/wearable-hrv/commit/8badf868308feafc8f984e749957bc653295d300).
- I have also changed and updated the RTD documentation, and also host the outputs of the overview notebook there: https://wearable-hrv.readthedocs.io/en/latest/
