### 2024-03-22 Aminsinichi/wearable-hrv#2 @Aminsinichi
Sure! Now, in all the notebook examples, as well as in the README and RTD, I have clarified what I mean by "criterion."
