### 2024-03-23 dc6246e
update import_data and chop_data functions to avoid error
