### 2024-03-23 755490c
update display_changes to avoid error
