+++ paper/paper.bib
@@ -171,6 +171,15 @@
   year={2010},
 }
 
+# pyHRV
+@inproceedings{Gomes2019,
+   author = {Gomes, Pedro and Margaritoff, Petra and Silva, Hugo},
+   booktitle = {Proc. Int'l Conf. on Electrical, Electronic and Computing Engineering (IcETRAN)},
+   pages = {822-828},
+   title = {{pyHRV: Development and evaluation of an open-source python toolbox for heart rate variability (HRV)}},
+   year = {2019}
+}
+
 
 # papers
 @article{hill_ethnic_2015,
@@ -322,3 +331,76 @@ Publisher: Nature Publishing Group},
   Publisher: Nature Publishing Group},
 	pages = {1--9},
 }
+
+@article{BRUTON200094,
+title = {Reliability: What is it, and how is it measured?},
+journal = {Physiotherapy},
+volume = {86},
+number = {2},
+pages = {94-99},
+year = {2000},
+issn = {0031-9406},
+doi = {https://doi.org/10.1016/S0031-9406(05)61211-4},
+url = {https://www.sciencedirect.com/science/article/pii/S0031940605612114},
+author = {Anne Bruton and Joy H Conway and Stephen T Holgate},
+keywords = {Reliability, measurement, quantitative measures, statistical method},
+abstract = {Summary
+Therapists regularly perform various measurements. How reliable these measurements are in themselves, and how reliable therapists are in using them, is clearly essential knowledge to help clinicians decide whether or not a particular measurement is of any value. The aim of this paper is to explain the nature of reliability, and to describe some of the commonly used estimates that attempt to quantify it. An understanding of reliability, and how it is estimated, will help therapists to make sense of their own clinical findings, and to interpret published studies. Although reliability is generally perceived as desirable, there is no firm definition as to the level of reliability required to reach clinical acceptability. As with hypothesis testing, statistically significant levels of reliability may not translate into clinically acceptable levels, so that some authors' claims about reliability may need to be interpreted with caution. Reliability is generally population specific, so that caution is also advised in making comparisons between studies. The current consensus is that no single estimate is sufficient to provide the full picture about reliability, and that different types of estimate should be used together.}
+}
+
+@article{d66d4fe8-8a27-315e-bbce-f72f450c6450,
+ ISSN = {00390526, 14679884},
+ URL = {http://www.jstor.org/stable/2987937},
+ abstract = {Methods of analysis used in the comparison of two methods of measurement are reviewed. The use of correlation, regression and the difference between means is criticized. A simple parametric approach is proposed based on analysis of variance and simple graphical methods.},
+ author = {D. G. Altman and J. M. Bland},
+ journal = {Journal of the Royal Statistical Society. Series D (The Statistician)},
+ number = {3},
+ pages = {307--317},
+ publisher = {[Royal Statistical Society, Wiley]},
+ title = {Measurement in Medicine: The Analysis of Method Comparison Studies},
+ urldate = {2024-03-26},
+ volume = {32},
+ year = {1983}
+}
+
+@article{MAKRIDAKIS1993527,
+title = {Accuracy measures: theoretical and practical concerns},
+journal = {International Journal of Forecasting},
+volume = {9},
+number = {4},
+pages = {527-529},
+year = {1993},
+issn = {0169-2070},
+doi = {https://doi.org/10.1016/0169-2070(93)90079-3},
+url = {https://www.sciencedirect.com/science/article/pii/0169207093900793},
+author = {Spyros Makridakis}
+}
+
+@article{Haghayegh_2020,
+doi = {10.1088/1361-6579/ab86d6},
[…]
