### 2024-03-26 Aminsinichi/wearable-hrv#10 @Aminsinichi
Thank you! I added a paragraph and situated the package among the current existing packages to demonstrate its added value. Additionally, I included a few classic references on papers about using validation methods and their interpretations, which are also utilized in the `wearablehrv` package.

I also made changes and additions to ensure it remains updated with the latest updates of the package.
