### 2024-06-29 #6240 @arfon
:wave: @richrobe @arianesasso @VBlesius – would any of you be willing to review this submission for JOSS? The submission under consideration is _Wearablehrv: A Python package for the validation of heart rate and heart rate variability in wearables_

Based on your experience, and past papers you have published in JOSS, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
