+++ wearablehrv/__init__.py
@@ -2,7 +2,7 @@
 #### A Python package designed for data preparation, pre-processing, feature exctraction, comparison, visualization, and statistical individual/group analysis of heart rate and heart rate variability outcome variables recorded from wearable devices that transmit raw interbeat intervals to a host device via Bluetooth.
 ############################ Wearablehrv ############################
 
-__version__ = "0.1.13"
+__version__ = "0.1.14"
 
 from . import individual
 from . import group
