### 2024-07-26 #6240 @Aminsinichi
@arfon - Thank you very much! I arranged the following:

- Edited the paper.md slightly and pushed it on GitHub.
- Created a final release (0.2.0)
- Archived it on Zenodo with the following DOI: https://doi.org/10.5281/zenodo.12917981

Let me know if anything else is required from my side.
