### 2024-08-23 #6240 @samhforbes
Hi @Aminsinichi can you please check whether the suggested missing DOI should be added, and the two invalid DOIs are corrected?
