### 2024-08-23 31af17d
update paper.bib to fix DOIs
