+++ paper/paper.bib
@@ -340,7 +340,7 @@ number = {2},
 pages = {94-99},
 year = {2000},
 issn = {0031-9406},
-doi = {https://doi.org/10.1016/S0031-9406(05)61211-4},
+doi = {10.1016/S0031-9406(05)61211-4},
 url = {https://www.sciencedirect.com/science/article/pii/S0031940605612114},
 author = {Anne Bruton and Joy H Conway and Stephen T Holgate},
 keywords = {Reliability, measurement, quantitative measures, statistical method},
@@ -360,7 +360,8 @@ Therapists regularly perform various measurements. How reliable these measuremen
  title = {Measurement in Medicine: The Analysis of Method Comparison Studies},
  urldate = {2024-03-26},
  volume = {32},
- year = {1983}
+ year = {1983}, 
+ doi = {10.2307/2987937}
 }
 
 @article{MAKRIDAKIS1993527,
@@ -371,7 +372,7 @@ number = {4},
 pages = {527-529},
 year = {1993},
 issn = {0169-2070},
-doi = {https://doi.org/10.1016/0169-2070(93)90079-3},
+doi = {10.1016/0169-2070(93)90079-3},
 url = {https://www.sciencedirect.com/science/article/pii/0169207093900793},
 author = {Spyros Makridakis}
 }
