### 2024-08-26 #6240 @arfon
@Aminsinichi – I found a tiny typo fix here: https://github.com/Aminsinichi/wearable-hrv/pull/11
