### 2024-08-26 #6240 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
